Add round trip tests for Any
During this test I discovered a bug in writeStringAP. I was specifying
the length as the length of characters in the strength which works when
the encoding using the same number of bytes as the original string.
Unfortunately this is not guaranteed to always be the case. The tests
for Any caught this bug as they generated strings with ‘random’ unicode
chars in them.

This also points out that I don’t have all of the round trip tests
written for all of the types.