Extract tag writers into Tag.Writer.hs
Delete BACnet.Tag.hs

This cleans up the modules a bit more. Especially the imports of
BACnet.Tag were very long because we had all of the imports needed for
reading and writing. I think this is cleaner.

(Although it might still be nice to bring back BACnet.Tag as a module
to export all the submodules. But only BACnet.Reader and BACnet.Writer
use it and they can just import what they need.)