Fix a problem with compiling/linking
Apparently something in the environment changed. I discovered
I could no longer run my code in ghci. It boiled down to not
being able to link to the unfoldNum function in
BACnet.Writer.UnfoldNum. I see that this wasn't being mentioned
in the cabal file so I added it. (Not sure if that was the fix,
or if the unregistering and cleaning and rebuilding fixed it)